Move StatsGranularityV4 extension to a separate file
