+++ WooCommerce/Classes/ViewRelated/Dashboard/Stats v4/StatsTimeRangeBarViewModel.swift
@@ -95,40 +95,3 @@ struct StatsTimeRangeBarViewModel: Equatable {
                                                 timezone: timezone)
     }
 }
-
-extension StatsGranularityV4 {
-    var displayText: String {
-        switch self {
-        case .daily:
-            NSLocalizedString(
-                "statsGranularityV4.daily",
-                value: "By day",
-                comment: "Display text for the daily granularity of store stats on the My Store screen"
-            )
-        case .hourly:
-            NSLocalizedString(
-                "statsGranularityV4.hourly",
-                value: "By hour",
-                comment: "Display text for the hourly granularity of store stats on the My Store screen"
-            )
-        case .weekly:
-            NSLocalizedString(
-                "statsGranularityV4.weekly",
-                value: "By week",
-                comment: "Display text for the weekly granularity of store stats on the My Store screen"
-            )
-        case .monthly:
-            NSLocalizedString(
-                "statsGranularityV4.monthly",
-                value: "By month",
-                comment: "Display text for the monthly granularity of store stats on the My Store screen"
-            )
-        case .yearly:
-            NSLocalizedString(
-                "statsGranularityV4.yearly",
-                value: "By year",
-                comment: "Display text for the yearly granularity of store stats on the My Store screen"
-            )
-        }
-    }
-}
